Added new option to delete_disk to preserve image
This patch adds a new feature to delete_disk() to tell
the rbd-target-api to preserve the backing rbd image when
calling the delete disk command.   This effectively keeps the
rbd image in the pool, but removes the definition of the
volume from the gateway.  This is a clean remove for
managing the iscsi gateway when the volume is no longer exported
to any clients.